// src/app/api/type.ts
export enum LLMErrReason {
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  MODEL_NOT_AVAILABLE = "MODEL_NOT_AVAILABLE",
  NOT_ENOUGH_BALANCE = "NOT_ENOUGH_BALANCE",
  UNKOWN = "UNKOWN",
  FAILED_TO_AUTH = "FAILED_TO_AUTH",
}

export type LLMResponse = {
  reason: string;
  message: string;
};

export interface AIStreamParserOptions {
  event?: string;
}

export interface AIStreamParser {
  (
    data: string,
    options: AIStreamParserOptions,
  ): string | void | { isText: false; content: string };
}

// https://github.com/openai/openai-node/blob/07b3504e1c40fd929f4aae1651b83afc19e3baf8/src/resources/chat/completions.ts#L28-L40
interface ChatCompletionChunk {
  id: string;
  choices: Array<ChatCompletionChunkChoice>;
  created: number;
  model: string;
  object: string;
}
// ...
interface Completion {
  /**
   * A unique identifier for the completion.
   */
  id: string;

  /**
   * The list of completion choices the model generated for the input prompt.
   */
  choices: Array<CompletionChoice>;

  /**
   * The Unix timestamp of when the completion was created.
   */
  created: number;

  /**
   * The model used for completion.
   */
  model: string;

  /**
   * The object type, which is always "text_completion"
   */
  object: string;

  /**
   * Usage statistics for the completion request.
   */
  usage?: CompletionUsage;
}
// ...
export type AsyncIterableOpenAIStreamReturnTypes =
  | AsyncIterable<ChatCompletionChunk>
  | AsyncIterable<Completion>
  | AsyncIterable<any>;

type ExtractType<T> = T extends AsyncIterable<infer U> ? U : never;

export type OpenAIStreamReturnTypes =
  ExtractType<AsyncIterableOpenAIStreamReturnTypes>;

/**
 * Custom parser for AIStream data.
 * @interface
 * @param {string} data - The data to be parsed.
 * @param {AIStreamParserOptions} options - The options for the parser.
 * @returns {string | void} The parsed data or void.
 */
export interface AIStreamParser {
  (
    data: string,
    options: AIStreamParserOptions,
  ): string | void | { isText: false; content: string };
}

export interface FunctionDefinition {
  name: string;
  description: string;
  parameters: {
    type: "object";
    properties: Record<
      string,
      {
        type: string;
        description: string;
        enum?: string[];
      }
    >;
    required: string[];
  };
}